+++ boards/TMC5271_eval.c
@@ -868,8 +868,8 @@ static uint32_t handleParameter(uint8_t readWrite, uint8_t motor, uint8_t type, 
 	case 211:
 		//ADC Scaling Resitors
 		if(readWrite == READ) {
-			int val2 = (HAL.IOs->config->isHigh(Pins.IREF_R2));
-			int val3 = (HAL.IOs->config->isHigh(Pins.IREF_R3));
+			uint8_t val2 = (HAL.IOs->config->isHigh(Pins.IREF_R2));
+			uint8_t val3 = (HAL.IOs->config->isHigh(Pins.IREF_R3));
 			if (val2 == 0 && val3 == 0){ //48k
 				*value = 0;
 			}
@@ -931,8 +931,8 @@ static uint32_t handleParameter(uint8_t readWrite, uint8_t motor, uint8_t type, 
 	case 214:
 		// ADCTemperatur Converted
 		if(readWrite == READ) {
-			int adc = TMC5271_FIELD_READ(motorToIC(motor), TMC5271_IOIN, TMC5271_ADC_TEMPERATURE_MASK, TMC5271_ADC_TEMPERATURE_SHIFT);
-			*value = (int)((2.03*adc)-259);
+			int32_t adc = TMC5271_FIELD_READ(motorToIC(motor), TMC5271_IOIN, TMC5271_ADC_TEMPERATURE_MASK, TMC5271_ADC_TEMPERATURE_SHIFT);
+			*value = (int32_t)((2.03*adc)-259);
 		} else if(readWrite == WRITE) {
 			errors |= TMC_ERROR_TYPE;
 		}
@@ -1189,7 +1189,7 @@ static void periodicJob(uint32_t tick)
 	if(!noRegResetnSLEEP)
 		{
 	//check if reset after nSLEEP to HIGH was performed
-	for(int motor = 0; motor < TMC5271_MOTORS; motor++)
+	for(uint8_t motor = 0; motor < TMC5271_MOTORS; motor++)
 		{
 			tmc5271_periodicJob(&TMC5271, tick);
 		}
