+++ tmc/BLDC_LandungsbrueckeV3.c
@@ -38,10 +38,10 @@ uint8_t  pwmEnabled       = 0;
 uint8_t  bbmTime          = 50;
 uint8_t  motorPolePairs   = 1;
 
-int targetAngle         = 0;
-int hallAngle           = 0;
+int32_t targetAngle         = 0;
+int32_t hallAngle           = 0;
 
-int actualHallVelocity = 0; // electrical RPM
+int32_t actualHallVelocity = 0; // electrical RPM
 
 volatile int32_t adcSamples[8] = { 0 };
 uint8_t adcSampleIndex = 0;
@@ -295,7 +295,7 @@ void BLDC_calibrateADCs()
 	// the adc phase selection
 	nvic_irq_disable(TIMER0_UP_TIMER9_IRQn);
 
-	for (int myadc = 0; myadc < adcCount; myadc++)
+	for (uint8_t myadc = 0; myadc < adcCount; myadc++)
 	{
 		adc = myadc % 3;
 		// Select the ADC phase for offset compensation
@@ -362,18 +362,18 @@ void TIMER0_UP_TIMER9_IRQHandler(void)
 {
 	if(timer_interrupt_flag_get(TIMER0, TIMER_INT_FLAG_UP) == SET)
 	{
-		static int commutationCounter = 0;
-		static int velocityCounter = 0;
+		static int32_t commutationCounter = 0;
+		static int32_t velocityCounter = 0;
 
-		static int lastHallAngle = 0;
-		static int hallAngleDiffAccu = 0;
+		static int32_t lastHallAngle = 0;
+		static int32_t hallAngleDiffAccu = 0;
 
 		// Measure the hall sensor
 		HallStates actualHallState = inputToHallState(HAL.IOs->config->isHigh(Pins.HALL_U), HAL.IOs->config->isHigh(Pins.HALL_V), HAL.IOs->config->isHigh(Pins.HALL_W));
 		hallAngle = hallStateToAngle(actualHallState);
 
 		// Calculate the hall angle difference
-		int hallAngleDiff = (hallAngle - lastHallAngle);         // [-300 ; +360)
+		int32_t hallAngleDiff = (hallAngle - lastHallAngle);         // [-300 ; +360)
 		hallAngleDiff     = (hallAngleDiff + 360) % 360;         // [   0 ; +360)
 		hallAngleDiff     = ((hallAngleDiff + 180) % 360) - 180; // [-180 ; +180)
 		lastHallAngle = hallAngle;
@@ -611,7 +611,7 @@ int16_t BLDC_getTargetPWM()
 	return targetPWM;
 }
 
-int BLDC_getMeasuredCurrent()
+int32_t BLDC_getMeasuredCurrent()
 {
 	int32_t sum = 0;
 	for (uint8_t i = 0; i < ARRAY_SIZE(adcSamples); i++)
@@ -655,12 +655,12 @@ uint16_t BLDC_getOpenloopStepTime()
 	return openloopStepTime;
 }
 
-int BLDC_getTargetAngle()
+int32_t BLDC_getTargetAngle()
 {
 	return targetAngle;
 }
 
-int BLDC_getHallAngle()
+int32_t BLDC_getHallAngle()
 {
 	return hallAngle;
 }
@@ -699,7 +699,7 @@ int32_t BLDC_getActualOpenloopVelocity()
 }
 
 // Velocity measured by hall in RPM
-int BLDC_getActualHallVelocity()
+int32_t BLDC_getActualHallVelocity()
 {
 	return actualHallVelocity / motorPolePairs;
 }
